requirements.txt
Flask==3.1.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
prometheus-client
python-dotenv